requirements.txt
accelerate
datasets
diffusers
torchvision